Main line
The main part of the code is where the mouse file is taken in as input from a source path and the result is outputted to another file. Here the first part uses only every 100th line from each file and the second part uses every line from each file to filter. There are methods to remove stop words and get the stop word list and get every 100th line.

In the main line, the code takes a predetermined list of files from the overall mouse corpus (E.g 1998-2008), filters stop words from each file and returns the result to an output location. It continues to do this until it has gone through each file in the file list.